refactor and added haptic feedback on tabs and add to watchlist
