+++ pages/Details.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RenderDetails from '../components/RenderDetails';
+import RenderDetails from '../components/RenderMovieDetails';
 
 const Details = ({ route, navigation }) => {
   const { id } = route.params;
